+++ server/src/Routes/Admin/ProductSales.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE DataKinds #-}
+{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE TypeOperators #-}
@@ -7,19 +8,26 @@ module Routes.Admin.ProductSales
     , productSalesRoutes
     ) where
 
-import Control.Monad.Trans (liftIO)
+import Control.Monad (unless)
+import Control.Monad.Trans (lift, liftIO)
 import Data.Aeson ((.=), (.:), ToJSON(..), FromJSON(..), object, withObject)
+import Data.Maybe (catMaybes)
 import Data.Monoid ((<>))
 import Data.Time (LocalTime, TimeZone, getCurrentTimeZone, localTimeToUTC)
-import Database.Persist (Entity(..), SelectOpt(Desc), selectList, insert)
+import Database.Persist
+    ( Entity(..), SelectOpt(Desc), Update, selectList, insert, get, update
+    )
 import Database.Persist.Sql (fromSqlKey)
-import Servant ((:<|>)(..), (:>), AuthProtect, Get, Post, JSON, ReqBody)
+import Servant
+    ( (:<|>)(..), (:>), AuthProtect, Capture, Get, Post, Patch, JSON, ReqBody
+    , err404
+    )
 
 import Auth (Cookied, WrappedAuthToken, withAdminCookie, validateAdminAndParameters)
 import Models
 import Models.Fields (Cents, LotSize)
-import Routes.Utils (parseLocalTime)
-import Server (App, AppSQL, runDB)
+import Routes.Utils (mapUpdate, parseLocalTime, parseMaybeLocalTime)
+import Server (App, AppSQL, runDB, serverError)
 import Validation (Validation(..))
 
 import qualified Data.Text as T
@@ -32,12 +40,16 @@ type ProductSalesAPI =
          "list" :> ProductSalesListRoute
     :<|> "new" :> ProductSalesNewDataRoute
     :<|> "new" :> ProductSalesNewRoute
+    :<|> "edit" :> ProductSalesEditDataRoute
+    :<|> "edit" :> ProductSaleEditRoute
 
 
 type ProductSalesRoutes =
          (WrappedAuthToken -> App (Cookied ProductSalesListData))
     :<|> (WrappedAuthToken -> App (Cookied ProductSalesNewData))
     :<|> (WrappedAuthToken -> ProductSalesNewParameters -> App (Cookied ProductSaleId))
+    :<|> (WrappedAuthToken -> ProductSaleId -> App (Cookied ProductSalesEditData))
+    :<|> (WrappedAuthToken -> ProductSaleEditParameters -> App (Cookied ()))
 
 
 productSalesRoutes :: ProductSalesRoutes
@@ -45,9 +57,27 @@ productSalesRoutes =
          productSalesListRoute
     :<|> productSalesNewDataRoute
     :<|> productSalesNewRoute
+    :<|> productSalesEditDataRoute
+    :<|> productSalesEditRoute
 
 
 -- COMMON
+
+newtype ProductSaleData =
+    ProductSaleData
+        { fromProductSaleData :: Entity ProductSale
+        } deriving (Show)
+
+instance ToJSON ProductSaleData where
+    toJSON ProductSaleData {..} =
+        let (Entity saleId ProductSale {..}) = fromProductSaleData in
+        object
+            [ "id" .= saleId
+            , "price" .= productSalePrice
+            , "variant" .= productSaleProductVariantId
+            , "start" .= productSaleStartDate
+            , "end" .= productSaleEndDate
+            ]
 
 data SaleVariantData =
     SaleVariantData
@@ -76,8 +106,8 @@ getProductsAndVariants =
         E.on $ v E.^. ProductVariantProductId E.==. p E.^. ProductId
         return (v, p)
 
-toVariantData :: Entity Product -> Entity ProductVariant -> SaleVariantData
-toVariantData (Entity _ p) (Entity vId v) =
+toVariantData :: Entity ProductVariant -> Entity Product -> SaleVariantData
+toVariantData (Entity vId v) (Entity _ p) =
     SaleVariantData
         { svdId = vId
         , svdName = productName p
@@ -97,7 +127,7 @@ type ProductSalesListRoute =
 
 data ProductSalesListData =
     ProductSalesListData
-        { psldSales :: [ListProductSales]
+        { psldSales :: [ProductSaleData]
         , psldVariants :: M.Map ProductVariantId SaleVariantData
         } deriving (Show)
 
@@ -108,33 +138,16 @@ instance ToJSON ProductSalesListData where
             , "variants" .= M.mapKeys (show . fromSqlKey) psldVariants
             ]
 
-newtype ListProductSales =
-    ListProductSales
-        { fromListProductSales :: Entity ProductSale
-        } deriving (Show)
-
-instance ToJSON ListProductSales where
-    toJSON ListProductSales {..} =
-        let (Entity saleId ProductSale {..}) = fromListProductSales in
-        object
-            [ "id" .= saleId
-            , "price" .= productSalePrice
-            , "variant" .= productSaleProductVariantId
-            , "start" .= productSaleStartDate
-            , "end" .= productSaleEndDate
-            ]
-
-
 productSalesListRoute :: WrappedAuthToken -> App (Cookied ProductSalesListData)
 productSalesListRoute = flip withAdminCookie $ \_ -> runDB $ do
-    sales <- map ListProductSales <$> selectList [] [Desc ProductSaleEndDate]
+    sales <- map ProductSaleData <$> selectList [] [Desc ProductSaleEndDate]
     ProductSalesListData sales . makeVariants <$> getProductsAndVariants
   where
     makeVariants :: [(Entity ProductVariant, Entity Product)] -> M.Map ProductVariantId SaleVariantData
     makeVariants =
         foldr
             (\(e@(Entity vId _), p) acc ->
-                M.insert vId (toVariantData p e) acc
+                M.insert vId (toVariantData e p) acc
             ) M.empty
 
 
@@ -155,7 +168,7 @@ instance ToJSON ProductSalesNewData where
 
 productSalesNewDataRoute :: WrappedAuthToken -> App (Cookied ProductSalesNewData)
 productSalesNewDataRoute = flip withAdminCookie $ \_ -> runDB $
-    ProductSalesNewData . map (\(v, p) -> toVariantData p v) <$> getProductsAndVariants
+    ProductSalesNewData . map (uncurry toVariantData) <$> getProductsAndVariants
 
 type ProductSalesNewRoute =
        AuthProtect "cookie-auth"
@@ -201,3 +214,100 @@ productSalesNewRoute = validateAdminAndParameters $ \_ params -> do
             , productSaleStartDate = localTimeToUTC timeZone psnpStart
             , productSaleEndDate = localTimeToUTC timeZone psnpEnd
             }
+
+
+-- EDIT
+
+type ProductSalesEditDataRoute =
+       AuthProtect "cookie-auth"
+    :> Capture "id" ProductSaleId
+    :> Get '[JSON] (Cookied ProductSalesEditData)
+
+data ProductSalesEditData =
+    ProductSalesEditData
+        { psedSale :: ProductSaleData
+        , psedVariants :: [SaleVariantData]
+        } deriving (Show)
+
+instance ToJSON ProductSalesEditData where
+    toJSON ProductSalesEditData {..} =
+        object
+            [ "sale" .= psedSale
+            , "variants" .= psedVariants
+            ]
+
+productSalesEditDataRoute :: WrappedAuthToken -> ProductSaleId -> App (Cookied ProductSalesEditData)
+productSalesEditDataRoute t saleId = withAdminCookie t $ \_ -> runDB $
+    get saleId >>= \case
+        Nothing ->
+            lift $ serverError err404
+        Just productSale -> do
+            variants <- map (uncurry toVariantData) <$> getProductsAndVariants
+            return ProductSalesEditData
+                { psedSale = ProductSaleData $ Entity saleId productSale
+                , psedVariants = variants
+                }
+
+
+type ProductSaleEditRoute =
+       AuthProtect "cookie-auth"
+    :> ReqBody '[JSON] ProductSaleEditParameters
+    :> Patch '[JSON] (Cookied ())
+
+data ProductSaleEditParameters =
+    ProductSaleEditParameters
+        { psepId :: ProductSaleId
+        , psepPrice :: Maybe Cents
+        , psepVariant :: Maybe ProductVariantId
+        , psepStart :: Maybe LocalTime
+        , psepEnd :: Maybe LocalTime
+        } deriving (Show)
+
+instance FromJSON ProductSaleEditParameters where
+    parseJSON = withObject "ProductSaleEditParameters" $ \v -> do
+        psepId <- v .: "id"
+        psepPrice <- v .: "price"
+        psepVariant <- v .: "variant"
+        psepStart <- v .: "start" >>= parseMaybeLocalTime
+        psepEnd <- v .: "end" >>= parseMaybeLocalTime
+        return ProductSaleEditParameters {..}
+
+instance Validation ProductSaleEditParameters where
+    validators ProductSaleEditParameters {..} = do
+        exists <- V.exists psepId
+        variantCheck <- case psepVariant of
+            Nothing -> return []
+            Just variantId -> do
+                variantExists <- V.exists variantId
+                return
+                    [ ( "variant"
+                      , [ ( "Could not find this Product in the database."
+                          , variantExists
+                          )
+                        ]
+                      )
+                    ]
+        return $
+            ( ""
+            , [ ( "Could not find this Product Sale in the database", exists )
+              ]
+            )
+            : variantCheck
+
+productSalesEditRoute :: WrappedAuthToken -> ProductSaleEditParameters -> App (Cookied ())
+productSalesEditRoute = validateAdminAndParameters $ \_ params -> do
+    timeZone <- liftIO getCurrentTimeZone
+    let updates = makeUpdates params timeZone
+    unless (null updates) $
+        runDB $ update (psepId params) updates
+  where
+    makeUpdates :: ProductSaleEditParameters -> TimeZone -> [Update ProductSale]
+    makeUpdates ProductSaleEditParameters {..} zone =
+        let mapUpdateUTC f v = mapUpdate f $ localTimeToUTC zone <$> v
+        in
+        catMaybes
+            [ mapUpdate ProductSalePrice psepPrice
+            , mapUpdate ProductSaleProductVariantId psepVariant
+            , mapUpdateUTC ProductSaleStartDate psepStart
+            , mapUpdateUTC ProductSaleEndDate psepEnd
+            ]